refactor(protocol): extract pending queue module and use explicit imports
Extract the ~540-line pending decryption queue subsystem from mod.rs into
a dedicated pending_queue.rs sub-module. This code has its own invariants,
metrics, and eviction policies — it belongs in its own file.

Replace `use super::*` wildcard imports in all protocol sub-modules with
explicit imports, making each module's dependencies self-documenting.
Sub-modules now import protocol-local items from `super::` and external
crate items directly from their source.
