+++ crates/offline-protocol/src/protocol/config_accessors.rs
@@ -1,6 +1,22 @@
 //! Configuration accessors, diagnostics, and service registration.
 
-use super::*;
+use super::{
+    lock_shared_state, OfflineProtocol, PendingQueueMetrics, ProtocolState,
+    MEDIA_TRANSFER_STALE_TIMEOUT_SECS,
+};
+use crate::file_transfer::FileTransferManager;
+use crate::{Error, ProtocolConfig, Result, TransportManager};
+use offline_protocol_core::{MessageId, ServiceDescriptor};
+use offline_protocol_mls::MlsManager;
+use offline_protocol_reliability::{
+    AckConfig, AckManager, Deduplicator, DeduplicatorConfig, DeduplicatorStats, RetryConfig,
+    RetryQueue,
+};
+use offline_protocol_router::DorsConfig;
+use offline_protocol_services::MeshServices;
+use std::sync::{Arc, RwLock};
+use std::time::Duration as StdDuration;
+use tracing::{error, warn};
 
 impl OfflineProtocol {
     /// Gets the current protocol state.
