+++ crates/offline-protocol/src/protocol/message_dispatch.rs
@@ -1,6 +1,20 @@
 //! Message dispatch handlers for internal protocol messages.
 
-use super::*;
+use super::{
+    internal_prefixes, lock_shared_state, ConnectionAcceptedPayload, ConnectionRequestPayload,
+    GroupCreatedPayload, GroupErrorPayload, GroupInfoPayload, GroupMemberAddedPayload,
+    GroupMemberRemovedPayload, GroupMessageReceivedPayload, InternalMessageResult,
+    KeyPackagePayload, OfflineProtocol, PresencePayload, ReadReceiptPayload, ReceivedKeyPackage,
+    TypingIndicatorPayload, UserGroupsPayload, MAX_READ_RECEIPT_IDS,
+};
+use crate::events::{DecryptionFailureCode, Event};
+use crate::mls_observability::{DecryptionFailureKind, MlsErrorCategory, MlsOperationContext};
+use crate::SessionStateError;
+use chrono::Utc;
+use offline_protocol_core::{Message, MessagePriority};
+use offline_protocol_mls::{EncryptedMessage, WelcomeMessage};
+use offline_protocol_services::ServiceAction;
+use tracing::{debug, info, warn};
 
 impl OfflineProtocol {
     /// Handles an incoming MLS key package message.
