+++ crates/offline-protocol/src/protocol/security.rs
@@ -1,7 +1,15 @@
 //! Control message signing, verification, and TOFU key management.
 
-use super::*;
-use offline_protocol_core::UserId;
+use super::{
+    base64_decode, base64_encode, storage_keys, InternalMessageResult, OfflineProtocol, TofuEntry,
+    CTRL_PK_META_KEY, CTRL_SIGN_DOMAIN, CTRL_SIG_META_KEY, DATA_PLANE_PREFIXES, INTERNAL_PREFIXES,
+    MAX_TOFU_PEERS, TOFU_MIN_EVICTION_AGE_MS,
+};
+use crate::events::Event;
+use crate::{Error, Result};
+use chrono::Utc;
+use offline_protocol_core::{Message, UserId};
+use tracing::{debug, error, info, warn};
 
 impl OfflineProtocol {
     /// Builds a canonical signing payload using length-prefixed encoding.
