+++ crates/offline-protocol/src/protocol/tests/mod.rs
@@ -1,12 +1,17 @@
 use super::*;
+use crate::constants::ACK_FOR_KEY;
+use crate::events::{DecryptionFailureCode, PresenceStatus};
 #[cfg(feature = "mls-observability")]
 use crate::mls_observability::MlsLifecycleEvent;
+use chrono::Duration as ChronoDuration;
+use offline_protocol_core::{AppId, MessagePriority, ServiceDescriptor, UserId};
 use offline_protocol_transport::{
     mock::MockTransport, Transport, TransportMetrics, TransportStatus, TransportType,
 };
 use std::sync::{Arc, Barrier, Mutex};
 use std::thread;
 use std::time::Duration;
+use std::time::Duration as StdDuration;
 
 pub(crate) fn create_test_config() -> ProtocolConfig {
     ProtocolConfig::new("test-app", "user123")
